fix(create): clear config cache after updating options
We need to do this otherwise cosmiconfig will return us the old, cached
options when we want to do our first install. This has only been causing
us obvious issues since the migration to zod because the validation
complains about the options that weren't defined when originally loading
the config, but it would likely have caused subtle bugs in the
installation process previously too.
